requirements.txt
# These libraries were used using python 3.10.8
#
# --- How create a virtual environment ---
# python -m venv venv
#
# Linux -  Mac systems
#     source venv/bin/activate
# Windows system using git bash or WSL:
#     source venv/Scripts/activate
#
# pip install --no-cache-dir -r requirements.txt

pyprojroot==0.2.0
ipykernel==6.17.0
matplotlib==3.6.0
google-auth==2.14.0
pandas==1.5.1
openpyxl==3.0.10
pandas-gbq==0.17.9
scikit-learn==1.1.3
psycopg2==2.9.4
seaborn==0.12.1
mysql-connector-python==8.0.31
